perf: extract Clock component to isolate 1Hz re-renders from StatusBar
